More Javadocs and extracted interface for RideService.
